+++ packages/form/src/TelephoneNumber/index.js
@@ -1,0 +1,1 @@
+export { default } from './TelephoneNumber';